+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Some checks (for ex. tags in constructed objects) now return specific errors (`InvalidTag`)
   instead of generic errors (`Verify`)
 - Refactor BerObject for parsing of tagged and optional values
+- Add method `as_bitslice()` to BerObject
 - Remove Copy trait from BerObjectHeader, copy is non-trivial and should be explicit
 - Fix the bug that caused OIDs longer than two subidentifiers which started by subidentifiers "0.0" ("itu-t recommenation") not to be decoded correctly
 - Implement the `as_u64` and `as_u32` methods for BerObjects with contents of type `BerObjectContent::BitString`.